Move diff validation test to test_operator_validation.py
